parameterize secret names
this commit adds logic to parameterize secret names for:

* judge server for both eval phases
* teacher server for sdg phase

the same fetch_secret() function is duplicated to ensure that we are not
passing secret data around as input/output parameters/artifacts. Doing
the latter would result in user secrets being stored in mlmd/object
store which we should avoid. In the future this logic will be replaced
with kfp built in secret mounting once it supports parameterization, a
lot of the duplicated logic will be removed.

We also perform rest requests against the host cluster because access to
kubernetes python package is not guaranteed.

Signed-off-by: Humair Khan <HumairAK@users.noreply.github.com>